+++ src/main.ts
@@ -3,6 +3,7 @@ import { pneumaPayoutsApi } from "./api";
 import { MODULE_ID } from "./constants";
 import { registerDiscordLinks } from "./discord-summary";
 import { registerPayoutLedger } from "./payout-ledger";
+import { ensurePayoutLog, registerPayoutLogSettings } from "./payout-log";
 import {
   hasInboxItemsForCurrentUser,
   openPayoutInbox,
@@ -29,6 +30,7 @@ Hooks.once("init", () => {
   registerPayoutJournalSettings();
   registerPayoutInboxSettings();
   registerPayoutDateSetting();
+  registerPayoutLogSettings();
 
   const module = game.modules.get(MODULE_ID);
   if (module) module.api = pneumaPayoutsApi;
@@ -36,6 +38,9 @@ Hooks.once("init", () => {
 
 Hooks.once("ready", () => {
   console.info(`${MODULE_ID} | Ready`);
-  if (game.user?.isGM) void ensurePayoutJournal();
+  if (game.user?.isGM) {
+    void ensurePayoutJournal();
+    void ensurePayoutLog();
+  }
   if (!game.user?.isGM && hasInboxItemsForCurrentUser()) openPayoutInbox();
 });